+++ R/googleAuthR_options.R
@@ -8,6 +8,7 @@
     googleAuthR.webapp.client_id = "201908948134-cjjs89cffh3k429vi7943ftpk3jg36ed.apps.googleusercontent.com",
     googleAuthR.webapp.client_secret = "mE7rHl0-iNtzyI1MQia-mg1o",
     googleAuthR.webapp.port = 1221,
+    googleAuthR.jsonlite.simplifyVector = TRUE,
     googleAuthR.scopes = c("https://www.googleapis.com/auth/webmasters",
                            "https://www.googleapis.com/auth/analytics",
                            "https://www.googleapis.com/auth/analytics.readonly",